+++ crates/nostr/src/prelude.rs
@@ -29,6 +29,7 @@ pub use crate::key::{self, *};
 pub use crate::message::{self, *};
 // NIPs
 pub use crate::nips::nip01::{self, *};
+pub use crate::nips::nip02::{self, *};
 #[cfg(feature = "nip04")]
 pub use crate::nips::nip04;
 #[cfg(all(feature = "std", feature = "nip05"))]