+++ CHANGELOG.md
@@ -40,6 +40,7 @@
 * nostr: change `Filter::custom_tag` value arg type ([Yuki Kishimoto])
 * nostr: rename `Filter::remove_custom_tag` to `Filter::remove_custom_tags` ([Yuki Kishimoto])
 * nostr: take a single filter per REQ and COUNT ([Yuki Kishimoto])
+* nostr: rename `contact` module to `nip02` ([Yuki Kishimoto])
 * pool: change `Relay::connect` method signature ([Yuki Kishimoto])
 * pool: change `Relay::disconnect` method signature ([Yuki Kishimoto])
 * pool: change `RelayPool::disconnect` method signature ([Yuki Kishimoto])